Update peer dependencies script
