+++ .eslintrc
@@ -1,5 +1,17 @@
 {
-  "extends": [
-    "./ts/node_modules/@kaliber/build/.eslintrc"
-  ]
+  "extends": "./node_modules/@kaliber/eslint-plugin/.eslintrc",
+  "rules": {
+    "@kaliber/no-relative-parent-import": "off"
+  },
+  "settings": {
+    "import/resolver": {
+      "node": {
+        "moduleDirectory": [
+          "node_modules",
+          "typescript/node_modules"
+        ]
+      },
+      "@kaliber/eslint-plugin/lib/absolute-path-resolver-plugin": null
+    }
+  }
 }
