+++ .stylelintrc
@@ -1,3 +1,4 @@
 {
-  "extends": "./ts/node_modules/@kaliber/build/.stylelintrc"
+  "ignoreFiles": ["node_modules/**/*", "**/*.js", "**/*.svg", "**/*.md"],
+  "extends": "./node_modules/@kaliber/stylelint-plugin/.stylelintrc"
 }
