--- .yarnrc
@@ -1,1 +1,0 @@
---*.modules-folder ts/node_modules
