+++ tsconfig.json
@@ -2,19 +2,31 @@
   "typeAcquisition": {
     "enable": true
   },
-  "include": ["./src/**/*", "./types/**/*"],
+  "include": [
+    "./scripts/**/*",
+    "./src/**/*",
+    "./types/**/*"
+  ],
   "compilerOptions": {
-    "moduleResolution": "node",
+    "module": "preserve",
+    "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "allowSyntheticDefaultImports": true,
-    "lib": ["dom", "esnext"],
+    "allowUmdGlobalAccess": true,
+    "lib": [
+      "dom",
+      "esnext"
+    ],
     "jsx": "react",
     "baseUrl": ".",
     "allowJs": true,
     "checkJs": true,
     "noEmit": true,
+    "downlevelIteration": true,
+    "maxNodeModuleJsDepth": 3,
     "paths": {
-      "*": ["ts/node_modules/*"]
+      "*": ["./node_modules/@types/*", "./typescript/node_modules/*"],
+      "sanity/*": ["./typescript/node_modules/sanity/lib/*"]
     }
   }
 }
